pipeline/requirements.txt
beautifulsoup4==4.9.3
bitarray==1.2.1
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.22
dataclasses==0.8
decorator==4.4.2
descartes==1.1.0
editdistance==0.5.3
filelock==3.0.12
gensim==4.0.1
idna==2.10
importlib-metadata==3.10.0
isodate==0.6.0
jellyfish==0.8.2
joblib==1.0.1
keybert==0.2.0
kiwisolver==1.3.1
matplotlib==3.3.4
mizani==0.7.3
mrakun==0.44
networkx==2.5.1
nltk==3.6.1
numpy==1.19.5
ordered-set==4.0.2
packaging==20.9
palettable==3.3.0
pandas==1.1.5
patsy==0.5.1
pdfminer==20191125
Pillow==8.2.0
plotnine==0.8.0
py3plex==0.94
pyaml==20.4.0
pycryptodome==3.10.1
pyparsing==2.4.7
pyspellchecker==0.6.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
rake-nltk==1.0.4
rdflib==5.0.0
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
segtok==1.5.10
selenium==3.141.0
sentence-transformers==1.0.4
sentencepiece==0.1.95
six==1.15.0
smart-open==5.0.0
soupsieve==2.2.1
statsmodels==0.12.2
tabulate==0.8.9
threadpoolctl==2.1.0
tika==1.24
tokenizers==0.10.2
torch==1.8.1
tqdm==4.60.0
transformers==4.5.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yake==0.4.6
zipp==3.4.1
